requirements.txt
opencv-python==4.5.4.58
numpy==1.19.5
Pillow==8.4.0
samsungtvws==2.2.2